Memory improvements, speed and sensitivity, and search algorithm near the working status. Want more? &:-)

+++ src/bio/pih/index/EncodedSubSequencesIndex.java
@@ -43,20 +43,20 @@ public interface EncodedSubSequencesIndex {
 	 * @throws BioException
 	 * @throws ValueOutOfBoundsException
 	 */
-	public abstract long[] getMatchingSubSequence(String subSequenceString) throws IllegalSymbolException, BioException, ValueOutOfBoundsException;
+	public abstract int[] getMatchingSubSequence(String subSequenceString) throws IllegalSymbolException, BioException, ValueOutOfBoundsException;
 
 	/**
 	 * @param subSequence
 	 * @return a list containing the {@link SubSequenceIndexInfo} encoded, use {@link SubSequenceIndexInfo} to decode it.
 	 * @throws ValueOutOfBoundsException
 	 */
-	public abstract long[] getMachingSubSequence(SymbolList subSequence) throws ValueOutOfBoundsException;
+	public abstract int[] getMachingSubSequence(SymbolList subSequence) throws ValueOutOfBoundsException;
 
 	/**
 	 * @param encodedSubSequence 
 	 * @return a list containing the {@link SubSequenceIndexInfo} encoded, use {@link SubSequenceIndexInfo} to decode it.
 	 */
-	public abstract long[] getMachingSubSequence(short encodedSubSequence);
+	public abstract int[] getMachingSubSequence(short encodedSubSequence);
 
 	/**
 	 * @return a string containing the status of the index.
